requirements.txt
Flask==3.0.0
supabase==2.3.3
httpx==0.25.2
python-dotenv==1.0.0
gotrue 2.12.0
email-validator==2.1.0.post1
gunicorn==21.2.0